pyproject.toml
[project]
name = "entra-signin-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "msal>=1.35.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
